feat(contrib): init tool upload-kktix-ticket-csv-to-bigquery.py
The tool is used to upload the ticket information exported from kktix to
bigquery, and pre-process the "raw" data to be more bigquery friendly in
column naming of tables.

It's dangerous to upload data by default. So, we use dry-run mode by default.

We would like to make the column names as much consistent as possible
across years, so we use some heuristic. We may need to maintain the
heuristic annually. Luckily and ideally, the annual maintanence will be
one-off.