fix: move airflow.db to sqlite/airflow.db
The way we define in docker-compose creates a "airflow.db" directory if the `airflow.db`
file is absent, causing Airflow to load the empty directory as a SQLite database.

By changing it to "sqlite" directory, it loads the content in it if presents and creates
a new one if not.